requirements.txt
python-dotenv
black
openai
anthropic
google.generativeai